backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
requests==2.32.3
pillow==11.0.0
schedule==1.2.2
feedparser==6.0.12
beautifulsoup4==4.13.3
python-dotenv
pytest==7.4.2
pytest-mock==3.11.4